requirements/htmx.txt
django-htmx
django-widget-tweaks==1.5.0
social-auth-core>=4.1
social-auth-app-django>=5.0
